updated colors, buttons and refactored editevent and profilepage
