+++ planit-react/public/manifest.json
