+++ planit-react/src/components/EventCard/EventCard.module.css
@@ -1,16 +1,21 @@
 .mainDiv {
-  border-radius: 30px;
+  border-radius: 15px;
   margin-top: 30px;
-  margin-bottom: 60px;
+  margin-bottom: 80px;
   margin-left: 160px;
   margin-right: 160px;
   padding: 20px 40px;
-  box-shadow: 5px 5px #e4e2e2cc;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
   border-style: inset;
   border-color: #e4e2e22d;
   overflow: hidden;
+  background-color: #f8f9fa;
 }
-
+.mainDiv:hover {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+  transform: scale(1.02);
+  transition: transform 0.1s ease-in-out;
+}
 .nameText {
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -23,7 +28,10 @@
 
 .info img {
   width: 200px;
+  height: 150px;
+  object-fit: cover;
   margin-right: 40px;
+  border-radius: 10%;
 }
 
 .descriptionDateContainer {
@@ -43,4 +51,17 @@
 .date {
   margin-left: 20px;
   width: 200px;
+}
+@media screen and (max-width: 768px) {
+  .mainDiv {
+    margin-left: 20px;
+    margin-right: 20px;
+  }
+  .description {
+    width: 100%;
+  }
+  .info img {
+    width: 100px;
+    height: 100px;
+  }
 }
