+++ planit-react/src/components/Footer/Footer.module.css
@@ -1,5 +1,6 @@
 .topMargin {
   margin-top: 100px;
+  background-color: rgb(248, 249, 250);
 }
 
 footer .follow {
