+++ planit-react/src/components/InputBox/InputBox.module.css
@@ -1,10 +1,3 @@
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  width: 100%;
-  background-color: white;
-}
 * {
   margin: 0;
   padding: 0;
@@ -75,12 +68,6 @@ button.transparant {
   border: 1px solid #fff;
   outline: none;
   transition: 0.2s;
-}
-
-:root {
-  --ITUgreen: #abcda4;
-
-  --SoftBlack: #1c1c1c;
 }
 
 .navbar {
