+++ planit-react/src/pages/EditEventPage/EditEventPage.module.css
@@ -1,0 +1,36 @@
+.formGroup {
+  text-align: center;
+  background-color: white;
+  font-family: "Gabarito", sans-serif;
+}
+.eventImagePreview {
+  display: block;
+  margin: 0 auto;
+  max-width: 50%;
+  height: auto;
+  margin-top: 10px;
+}
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-family: "Gabarito", sans-serif;
+}
+.form label,
+.form input,
+.form textarea,
+.form select {
+  display: block;
+}
+input[type="text"],
+input[type="date"],
+input[type="time"],
+textarea {
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  width: 100%;
+  font-size: 16px;
+  max-width: 500px;
+  border-radius: 20px;
+}
