+++ planit-react/src/pages/FeedPage/FeedPage.module.css
@@ -42,7 +42,7 @@
   position: absolute;
   left: 30px;
   top: 90px;
-  background-color: #ffffff;
+  mix-blend-mode: color-burn;
   color: rgb(97, 97, 97);
   border: none;
 }
